Support long CRC values under `dart2js` environment.
This is a big change. The most important ones are:

1. Instead of yielding an `int` as CRC value, it is now abstracted in a
`CrcValue` object so that we can use `BigInt` or `int` depending on the
width of the CRC value.

2. Low level primitives are split into `_js` and `_vm` files so that we
can deal with different integer bit-widths. Under JavaScript, the
integral shifts and bitwise operators only work with 32-bit integers.
The VM uses full 64-bit integers.

3. Similarly, a catalog of CRC functions are also split into two smaller
parts (plus a common shared collection of smaller sized CRCs). We do
this to optimize for the VM environment by taking advantage of 64-bit
integers, and fallback to the slower implementation with `BigInt` in the
JavaScript environment. It is a compilation error to write, for example,
0xFFFFFFFFFFFFFFFF when the code is under `dart2js`.